requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_openai
langchain-google-genai
langchain_experimental
streamlit 
openaiopenai 
ragas
tqdm
pandas
numpy

-e .